Move some source under common (upstream D70328)
