+++ openmp/libomptarget/deviceRTLs/amdgcn/src/omptarget-nvptx.h
@@ -20,9 +20,9 @@
 
 // local includes
 #include "target_impl.h"
-#include "debug.h"     // debug
+#include "common/debug.h"     // debug
 #include "interface.h" // interfaces with omp, compiler, and user
-#include "state-queue.h"
+#include "common/state-queue.h"
 #include "support.h"
 
 #define OMPTARGET_NVPTX_VERSION 1.1
