+++ openmp/libomptarget/deviceRTLs/amdgcn/src/support.cu
@@ -11,7 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "support.h"
-#include "debug.h"
+#include "common/debug.h"
 #include "omptarget-nvptx.h"
 
 ////////////////////////////////////////////////////////////////////////////////
